Fix datatype of temperature in workspace data entry.

+++ src/Mcce.SmartOffice.Api/Models/WorkspaceDataEntryModel.cs
@@ -6,11 +6,11 @@
 
         public DateTime Timestamp { get; set; }
 
-        public double Temperature { get; set; }
+        public float Temperature { get; set; }
 
-        public double Humidity { get; set; }
+        public float Humidity { get; set; }
 
-        public double Co2Level { get; set; }
+        public float Co2Level { get; set; }
 
         public int Wei { get; set; }
     }
